requirements.txt
streamlit==1.18.0
tensorflow==2.10.1
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0
altair<5